ShowPatchAction: Correct is_applicable()
Earlier, is_applicable() was not very smart and just checked
if the diff existed. Now, we check if the diff has a conflict.
If there is a conflict, this action is not applicable.
We use a similar condition to ApplyPatchAction, as when the
patch is viewable, it should also be applicable.

Fixes https://github.com/coala-analyzer/coala/issues/1235